# public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'Webapp API'
  version: '1.0'
paths:
  /api/v1/users:
    get:
      tags:
        - User
      summary: 'Returns a list of users'
      description: 'Returns a object of users'
      operationId: b343fe695d1beaa03b7deb8449dd1b7c
      responses:
        '200':
          description: 'A list with users'
      security:
        -
          apiAuth: []
  '/api/v1/users/{id}':
    post:
      tags:
        - user
      summary: 'Returns a single user'
      description: 'Returns a single user from ID'
      operationId: 37d7b39f83f01acf3c56f8242ad1cb16
      parameters:
        -
          name: id
          in: path
          description: 'User ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'A list with users'
      security:
        -
          apiAuth: []
  /api/v1/login:
    post:
      tags:
        - Login
      summary: Login
      operationId: login
      parameters:
        -
          name: email
          in: query
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json: {  }
        '401':
          description: Unauthenticated
        '400':
          description: 'Bad Request'
        '404':
          description: 'not found'
        '403':
          description: Forbidden
  /api/v1/register:
    post:
      tags:
        - Register
      summary: Register
      operationId: register
      parameters:
        -
          name: email
          in: query
          required: true
          schema:
            type: string
        -
          name: name
          in: query
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json: {  }
        '401':
          description: Unauthenticated
        '400':
          description: 'Bad Request'
        '404':
          description: 'not found'
        '403':
          description: Forbidden
components:
  securitySchemes:
    apiAuth:
      type: http
      name: bearerAuth
      in: header
      bearerFormat: JWT
      scheme: bearer